# Conflicts:
#	Backend/services/core/src/main/java/org/edu_sharing/repository/server/MCAlfrescoAPIClient.java
#	Backend/services/core/src/main/java/org/edu_sharing/repository/server/importer/ExcelLOMImporter.java
#	Backend/services/core/src/main/java/org/edu_sharing/service/clientutils/ClientUtilsService.java
#	Backend/services/core/src/main/java/org/edu_sharing/service/handleservice/HandleServiceImpl.java
#	Backend/services/core/src/main/java/org/edu_sharing/service/nodeservice/NodeServiceImpl.java
#	Frontend/projects/edu-sharing-ui/src/lib/node-entries/node-entries-wrapper.component.ts
#	Frontend/src/app/core-module
#	Frontend/src/app/pages/collections-page/collection-new/collection-new.component.ts
#	Frontend/src/app/pages/render-page/render-page.component.ts

+++ Backend/services/core/src/main/java/org/edu_sharing/restservices/NodeDao.java
@@ -223,7 +223,13 @@ public class NodeDao {
                     .forEach(defaultProps::remove);
 
             for (String prop : defaultProps) {
-                if (!props.containsKey(prop) && CCConstants.getValidGlobalName(prop) != null) {
+                if (!props.containsKey(prop) && CCConstants.getValidGlobalName(prop) != null &&
+                    // protected publish props
+                    !Arrays.asList(
+                            CCConstants.CCM_PROP_IO_PUBLISHED_ORIGINAL, CCConstants.CCM_PROP_IO_PUBLISHED_MODE,
+                            CCConstants.CCM_PROP_PUBLISHED_HANDLE_ID, CCConstants.CCM_PROP_IO_PUBLISHED_DATE
+                    ).contains(CCConstants.getValidGlobalName(prop))
+                ) {
                     // delete removed properties
                     nodeService.removeProperty(getStoreProtocol(), getStoreIdentifier(), getId(), CCConstants.getValidGlobalName(prop));
                 }
